Initial Commit Project 1
This is my initial commit for my Compiler project 1. I am not going to
lie i have been working on this for a while and i have a lot more than i
am committing now, but i didnt want to get minus infinity so i am
starting to commit by increments of my project. I also wanted to restart
a bit because my code is all messed up and ugly and all over the place.
so essentially im starting from scratch, but i have much of the code i
just need to reformat it and create better variable names and add better
comments. From the labs and segmenting project one into many little
projects and tasks i have much of the code, but i need to put it all
together to make it all work. I am also using the example of javascript
on the site as a reference and to get some ideas on how to get things to
work.