Committing my initial .css file
here is my basic .css file that i have linked to my page for styling
purposes. As i said this is initial and i will be adding to it when i
get things up and running. It will look very pretty when i add more and
try to make everything a little more visually appealing.