func tamessage, func parse, func chechToken and a few other
this is the initial commit for function parse, function parse program
and func parse block. Also i have created a very large function,
function checkToken. I have worked on this for a while, and have had
some trouble and spent a fair amount of time, but it is coming right
along. I have been in here since the library opened doing lots of
work... but now there are attractive females all around me and they are
talking and eating.... i smell smartfood... the distinctive smell of the
white sharp cheddar...smells very  synthetic in a way, but it still
makes me very hungry.  Luckily i have moved away from the distracting
people who use this library as a social club and i am hiding in the
corner from the smells of food. They should have people who walk around
here selling food or someone should create an app or online service to
order food in the lib and then it gets delivered to them and payment can
be cash, card, and or marist money.  i would use this a lot so i dont
have to be distracted and waste the time of walking to the cab or caf. i
may do this someday.... food for thought... no pun intended...or was it.
If you have read this far Alan, i think we should play golf sometime.
that would be cool.